Change body background colour and max width
when the page content is less than the viewport height, the body background should be light, the same background as the article. But when the viewport is wider than the max value, the background should be black/dark. By setting max-width on the body we can have vertical space fill with light colour and horizontal space fill with dark colour.